@@ -1,7 +1,10 @@
 package tech.yash;
 
-import org.junit.Before;
-import org.junit.Test;
+//import org.junit.Before;
+//import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.junit.rules.ExpectedException;
 
 import static org.junit.Assert.assertEquals;
@@ -12,7 +15,7 @@ public class NumberTest {
     Number number2;
     Number negative;
 
-    @Before
+    @BeforeEach
     public void setup(){
         number = new Number(67);
         number1 = new Number(25);
@@ -24,31 +27,38 @@ public class NumberTest {
 
     @Test
     public void primeTest1(){
-        assertEquals("isPrime Error 67 !", true, number.isPrime());
-        assertEquals("isPrime Error 25 !", false, number1.isPrime());
+        Assertions.assertEquals(true, number.isPrime(),"isPrime Error 67 !");
+        Assertions.assertEquals(false, number1.isPrime(), "isPrime Error 25 !");
 
     }
 
-    @Test(expected = IllegalArgumentException.class)
+    @Test
     public void primeTest_NonFunctional(){
 //        Number negative = new Number(-965);
 
-        number2.isPrime();
+        Assertions.assertThrows(IllegalArgumentException.class, ()->{
+            number2.isPrime();
+
+        });
     }
 
-    @Test(expected = IllegalArgumentException.class)
+    @Test
     public void primeTest_NonFunctional2(){
 //        Number number = new Number(0);
 
-        negative.isPrime();
+        Assertions.assertThrows(IllegalArgumentException.class, ()->{
+            negative.isPrime();
+
+
+        });
     }
 
     @Test
     public void armstrongTest1(){
         Number a = new Number(371);
         Number b = new Number(372);
-        assertEquals("isArmstrong Error 371 !", true, a.isArmstrong());
-        assertEquals("isArmstrong Error 372 !", false, b.isArmstrong());
+        Assertions.assertEquals(true, a.isArmstrong(), "isArmstrong Error 371 !");
+        Assertions.assertEquals(false, b.isArmstrong(),"isArmstrong Error 372 !");
 
     }
 
@@ -56,23 +66,31 @@ public class NumberTest {
     public void palindromeTest1(){
         Number a = new Number(110011);
         Number b = new Number(120011);
-        assertEquals("isPalindrome Error 110011 !", true, a.isPalindrome());
-        assertEquals("isPalindrome Error 120011 !", false, b.isPalindrome());
+        Assertions.assertEquals(true, a.isPalindrome(), "isPalindrome Error 110011 !");
+        Assertions.assertEquals(false, b.isPalindrome(), "isPalindrome Error 120011 !");
 
     }
 
-    @Test(expected = IllegalArgumentException.class)
+    @Test
     public void armstrongTest_NonFunctional2(){
 //        Number number = new Number(0);
         Number a = new Number(-965);
-        a.isArmstrong();
+
+        Assertions.assertThrows(IllegalArgumentException.class, ()->{
+            a.isArmstrong();
+
+        });
     }
 
-    @Test(expected = IllegalArgumentException.class)
+    @Test
     public void palindromeTest_NonFunctional2(){
 //        Number number = new Number(0);
         Number a = new Number(-965);
-        a.isPalindrome();
+
+        Assertions.assertThrows(IllegalArgumentException.class, ()->{
+            a.isPalindrome();
+
+        });
     }
 
 }
